Simplify and reduce duplication in PingbackTest
MediaWikiTestCaseTrait provides a helper method,
createNoOpMock, that covers the case of wanting
to assert that no methods or no methods except
those in a list get called.

Add a helper method for the repeated logic
of creating cache and config, and creating
and running the pingback.

Merge the two test cases for when a pingback
should actually be sent (no prior pingback in the
database, or the prior pingback is from more than a
month ago) to use a data provider.

Rename variables to be sligthly clearer.

Change-Id: I25daf4d78b7bea12a6bc68ace3827985e435d21d