WatchedItemStore: use UserFactory::newFromUserIdentity()
Currently, we have a UserIdentity, we retrieve the user id,
and then we create a User object based on the id. Instead,
create the User object based on the UserIdentity directly,
so that if the UserIdentity happens to already be a full User
object no change is needed.

This allows us to simplify WatchedItemStoreUnitTest by
not needing to tailor the mock UserFactory to the specific
user ids that we are using in each test case.

Change-Id: I52918b4f4c7d44086ea4fcdabdbfe14197c06373