BlockErrorFormatter: clean up handling of link to blocker's user page
Current code fetches a User based on their id, and then retrieves
a Title object for that User's user page via User::getUserPage().
That Title object is then used to retrieve the text and prefixed text.

Instead, retrieve the UserIdentity from the block directly,
and then create the PageReference manually. Use TitleFormatter to retrieve
the text and prefixed text.

Change-Id: Ic44bfb882523595566031a0fd788be4f15ef19e9