resourceloader: Allow require() on private files when under test
This is compatible with Node.js. It remains disabled in production
as module modules may not depend on internals of other modules, but this
can make things a lot simpler for testing.

The main use cases is to ease testing of private classes, without having
to conditionally export them "for tests only" from the public bundle
entry point (e.g. index.js). Instead, the internal export from the
private file can now be imported directly in tests.

This also eases mocking and stubbing of methods of those private classes
via Sinon. You can now use beforeEach() in a test to import and then
stub part of it via Sinon.

== Regex change ==

Allowing importing of paths like `modulename/foo.js` would, under the
previous path regex, introduce an ambiguity. Since change 229b11c6256
(I550d900e0b78e6bbc for T284707) a few months back, ResourceLoader
supports namespaced module names like `@vue/example`.

To resolve this ambiguity, we can use the same specification as
Node.js, which requires that at-names contain exactly 1 slash. In the
case of Node.js, they came at it from the inverse perspective, but
arriving at the same conclusion: They used to disallow "@" at, and
made it legal together with a slash. In our case, we used to disallow
slash "/", and we now make that legal if after an "@".

The Node.js equivalent to this is doucmented at
https://nodejs.org/docs/v20.1.0/api/modules.html#all-together
and implemented at
https://github.com/nodejs/node/blob/v20.1.0/lib/internal/modules/cjs/loader.js#L554-L560

Bug: T250045
Change-Id: I9fca9fdf9b7623b1a759a30f52872bdc269493b7