SkinTemplate: Factor out SkinTemplateNavigation hooks
Allow skins to run their own transformations on the result of
the SkinTemplateNavigation hooks

Note this makes a slight behavioural change to how the
SkinTemplateNavigation operates - it now uses the complete data set
for menus which includes variants. Previously
the Universal hook was the only way to modify this. We have
been discouraging use of the SkinTemplateNavigation hook and this
will not impact any known users.

Bug: T287533
Depends-On: I8969296c64875d58e3c7ea95f6b567916f82d8a7
Change-Id: Iec99be92378dc1a43f39428caffc03e1f594a927