UserNameUtils: cache message formatting for reserved usernames
We already cache the hook result to avoid running the hook more
than once, but we can also store the results of the ITextFormatter
calls to avoid repeating them.

This then allows us to use the simpler in_array function instead
of manually looping through the list of reserved usernames and
checking them individually.

Change-Id: I86fe1ecd7830952686118649df2f5810676248f3