maintenance: Remove LanguageFactory dependency from Benchmarker.php
By not pulling in the entire Language service tree, benchmarks become
standalone CLI scripts that work even when MW is not installed.

This is useful for benchmarks specifically, as those are generally
better run outside VMs one may have for developing MW, so all one has
to do is move or comment-out LocalSettings.php and then run the
benchmark in question.

Change-Id: Icfc0a818d307bce805da45c8304cfd6d3b810768