Build the page deletion form in DeleteAction, not Article
This is the first patch in that direction, and for now, the code is
**only moved** with minimal changes, to keep the diff small and make it
easier to review. More patches will be sent to clean up this code,
before moving parts of it to a separate service.

Article::delete() is being removed since there's going to be no
reasonable way to maintain BC, and it's also unused (which I think is
expected, given how many things that method would do).

The logic for varying the form depending on the page type is also being
moved to DeleteAction, which means there's currently no way for
extensions to customize that (and I couldn't find any extension doing
that at the moment). Should we need this behaviour, we could add a hook to
DeleteAction to allow altering the form.

Bug: T288282
Change-Id: Iadb8137d728f4fd4e749a1052f8aef4bab11ec09