Fix mw.util.getUrl when given empty title with fragment
In PHP, Title::getLinkURL() and related methods will make a
fragment-only link in this case (and ignore the given query
parameters). In JS we previously generated a link to the main page.
Change the JS code to match. Also add a unit test for this in PHP.

This also affects mw.Title#getUrl and the parsing of messages using
mediawiki.jqueryMsg.

Bug: T288415
Change-Id: I4aeba3409581f6a1d49a17e823ec83d8b76b844f