WebRequest: Document getVal caveat and recommend getText or getRawVal
Prior art in I8ce1fa31f, Id37619cead, I9db7b3f36f, If219463d80a.

Start to nudge future development in a better direction by starting
to discourage getVal() since it is an odd hybrid after all, that
for the most part could probably be replaced by getRawVal, except
that there is a handful of really critical getVal() calls for
"title" in many places that are difficult to track down, so for now
we're probably stuck with this.

Change-Id: I18767cd809f67bde6784bda5e3c85974ea2e7205