benchmarks: Remove some old and trivial benchmarks
These are mainly testing PHP rather MW code, so there isn't so much
to iterate or learn from these periodically.

As a general data point for what to pick one would presumably write a
micro bench from scratch for any of the thousands of possible things
one could think of varations for in the space of built-ins and such.

As a general benchmark for changing server or PHP configuration,
these would not be represenative either way, we'd likely use full
HTTP requests or something like the higher-level Parse and Tidy
benchmarks.

Change-Id: I4e98ae394a49c1f4a247b8d1e79a2ae5f7022d05