Move some methods from FileDeleteForm to a subclass of DeleteAction
This is where the UI logic is supposed to live. The remaining methods in
FileDeleteForm will later be deprecated in favour of a dedicated command
object.

This patch tries to move the code to the new class with minimal changes,
with cleanup left to do for later.

Also change return typehints and docs for getFile(): WikiFilePage::loadFile()
is guaranteed to load a file. The "|bool" would confuse PHPStorm and
phan.

Note that FileDeleteForm is unused everywhere, except for some static
methods not touched by this patch.

Bug: T288282
Change-Id: I4fa1dfa096cd2ae7a58067764b12a275b1cce1d5