DeleteAction: allow easy override of form messages
The two forms (article delete vs file delete) are basically identical,
but use different messages. Even if some file-specific messages might
possibly be deleted in favour of the base ones, it's still hard to
generate the form fields from a single place. Thus, this change
introduces a couple of methods to allow overriding the messages.

Bug: T288282
Change-Id: I1c26b24c16cb6ab6840edaca4bce0d582518b95d