Build the page deletion form in DeleteAction, not Article
This is the first patch in that direction, and for now, the code is
**only moved** with minimal changes, to keep the diff small and make it
easier to review. More patches will be sent to clean up this code,
before moving parts of it to a separate service.

Article::delete() and friends were hard-deprecated, given the lack of
callers according to codesearch (searched in "Everything").

The logic for varying the form depending on the page type is also being
moved to DeleteAction, which means there's currently no way for
extensions to customize that (and I couldn't find any extension doing
that at the moment). Should we need this behaviour, we could add a hook to
DeleteAction to allow altering the form.

Bug: T288282

Change-Id: I34525cf3e7c9a82032101b8c2e7b14d6dff59092