mediawiki.jqueryMsg: Reorder more code for readability
Some variables don't need to be declared at the top of the
relevant scope if they are only used after they are defined,
since we no longer enable the eslint rule vars-on-top. Move
the declarations to where they are defined.

Also, in mw.jqueryMsg.HtmlEmitter.emit replace a temporary
return variable with direct early returns.

Should be a no-op in terms of functionality.

Change-Id: I8c036c5305a99b93140a8d02faafa7bed928c0b6