mediawiki.jqueryMsg.js: remove setParserDefaults() deep parameter
The method is documented as being used in
ResourceLoaderJqueryMsgModule, but that was removed. The only
existing callers are in qunit tests - the base test runner calls
mw.jqueryMsg.setParserDefaults to update the reference to the
messages to use, and the jqueryMsg tests call the method in
order to test it. Since the base test runner doesn't use the `deep`
option, its only being invoked in the tests for the functionality -
remove the functionality and the tests for it.

Follow-up: I203d4e3ecdeeeb16729eba2dcf40d11a41d2e582
Change-Id: I28a28175dd295f631d63034e98481c357e0189ab