Make DeleteAction and FileDeleteAction share showForm
Use getTitle() in FileDeleteAction, it's the same as using the old
$this->file.

Use a single showForm method for both classes, and a separate method to
add handy links. When switching to FormAction, the former will be
replaced by getFormFields(), the latter by postText(). However, before
doing that, the following things should be addressed:
- The onArticleConfirmDelete hook, which allows changing the reason.
  This hook should be replaced in the long term, but for now it might be
  sufficient to add it to the common UI logic and only run it for
  non-file pages.
- The horrible $hasHistory hack. This is going to be a tad more
  complicated because we need to remove it from WikiPage and
  ContentHandler, and replace it with something else.

Bug: T288282
Change-Id: I965c32457b1426ece9781221a95337d295b9e1a2