objectcache: Ignore "failed to set" warning from php-memcached
This can sometimes be emitted from setMulti() since php-memcached 3.0.0
if a write failure happened. This is redundant with the error message
already obtained from getResultMessage() via checkResult(), which
are written to the 'memcached' or 'objectcache' log channels by default.

The warning is not actionable, and other write methods such as set(),
add(), and deleteMulti() don't emit such redundant warning for the
same kind of issue. Suppress it as such to reduce noise in production
error logs.

Bug: T251450
Change-Id: I4ce6c2aed8688a6c051655574e9a0bae637ac11e