resourceloader: Define mw.loader in its own JS file
Remove primary IIFE when setting mw.loader, which reduces
indentation by an extra level. Futher cleanup will be done
to the code for mw.loader.store in a follow-up.

The fnv132() function, which is only used in the loader, was
among the extra bits that were moved.

Should have no effect on the functionality.

Change-Id: Ie475879727e4170eb47f3fc78048379c7190844d