In Rest\ResponseFactory, substitute invalid UTF-8
Most of the time this is due to garbage user input, and MediaWiki's
normal handling of user input strings is to normalize them, not to
throw.

There are some hypothetical cases where a developer might want to see an
exception, rather than passing garbage to the client and letting the
client fail, but normalization is a defensible policy and works within
the existing interfaces.

Bug: T289597
Change-Id: Ida5ee72c10d1c3499b5b17594c5675db7cd38335