Add mw-list-item class to all menu list items created by makeListItem
When working on multiple skins its been confusing to know how the lists
are created from the HTML. In Vector we've also resorted to some overly
complex styling rules to style `li` elements. It's proposed a mw-list-item
class is added to simplify Vector styling and make it easier to understand
how the HTML is generated.

Change-Id: I67731fe59743620d339ffcca3ffc3b1f127ce3e7