rdbms: Change getRecordedTransactionLagStatus() to use time from lag check
Previously, it created a new array each time and took the 'lag' value
from the `getApproximateLagStatus()` call together with a time from
`trxTimestamp()` which is set assigned earlier in `begin()`.

The `getApproximateLagStatus()` method already returns a timestamp that
seems like it should be used instead, and would be less surprising if
we did.

I believe this would in practice slightly reduce the amount of lag
being reported, because we no longer count the time ellapsed between
the trxTimestamp and trxReplicaLagStatus assignment as additional
session lag. If this were pure PHP code, then that would be negligible,
but my guess is that it is not negligible because there is a network
roundtrip between these points for the lag query itself.

Given that getApproximateLagStatus() always assigns this after the
query, I assume that is what we prefer. Though, if we want to be more
pessimistic, we can change that method to obtain the time before calling
getLag().

Change-Id: Ibe3077eb7c1bcf51fc95623ec920ba1b146dfd2b