specials: Avoid Status/StatusValue mismatch in SpecialEmailUser
IEmailer::send is documented to return StatusValue,
but SpecialEmailUser::submit is documented to return Status.
The real implementation of the service in Emailer class is using
UserMailer::send, which returns Status. This is only a fix to make phan
happy and avoid issues if another service is used in the future.

Change-Id: I0d1a5560aa74ff5948d660bb406ee31d1cd4e6fb