mw.loader: simplify setAndPropagate(), inline requestPropagation()
setAndPropagate: simplify conditional handling of states. Instead
of checking if the state is 'loaded', 'ready', 'error', or 'missing',
and if so adding some extra conditional code for the states 'ready',
'error', and 'missing' and then calling requestPropagation(), just use
an early return if its not 'loaded', within an `else if` block that is
only if the state isn't one of the other three.

requestPropagation is only called in a single place, setAndPropagate,
and its a very simple method - merge it into the caller.

Should be a no-op in terms of functionality.

Follows-up I499ae3f095 (6a83163b1cb), which created these functions.

Change-Id: I9dd6994e55f4a93f17e4424c2b74648950e25a66