mw.loader: Simplify callback in nestedAddScript()
nestedAddScript() gets called once with a callback parameter,
and then calls itself recursively until all scrips are processed,
at which point it invokes that callback. Since there is only one
caller, we know exactly what the callback will be, and can just
call it directly instead of passing it as a parameter.

Should be a no-op in terms of functionality.

Change-Id: Id7d301bbbd95c4d9671d6a141552b463626289fd