resourceloader: Remove unneeded `undefined` checks in enqueue()
In enqueue(), instead of checking if the `ready` and `error` callbacks
are not `undefined`, just check if they are truthy.

This is a private method, called in three ways:

* With explicit undefined values.
* With our own inline or referenced functions.
* In mw.loader.using() with $.Deferred#reject,
  which is also known to be a function.

As such, today there are no falsey values other than undefined possible
here and this change does not alter behaviour.

== Background

These checks were added in 52b1cc3f530 (r112665), and existed in the
form of an arguments.length check before that. There were two important
differences compared to today:

1. The jobs object was used to control propagating state changes.
   It was important to add a job even if without ready/error callback.
   This was implemented by passing `null` when a job is needed,
   and no parameter (undefined) when no job is needed.

2. At this time, mw.loader.using() did not yet use $.Deferred for
   the user-supplied ready and error callback. Instead, these got passed
   directly into our job system. This was changed in 07010d0b17.

The ability to create a "propagate job without callbacks" was introduced
by accident as part of commit 47f9cf8bd58 (r110592) where a fourth
parameter to enqueue() was created and called with null ready/error
without changing the `arguments.length` check.

The ability was removed again in commit 62308668c34 (I6428e733899f7).

Change-Id: Iefa115986e5f60c79e16deb808dc5e85204e11a3