resourceloader: Unhoist more variables in mw.loader
Declare variables closer to where they are used, now that we no longer
enforce that all variables need to be declared at the top. In many cases
this should shave off a few characters from the output size, because
adding an extra variable declaration adds 4 characters (`var `), and
declaring the variable at the top before its being assigned is usually
more than 4 characters (comma, plus the length of the variable name).

In some cases, the variable could be removed entirely since it was
only used once.

Should be a no-op in terms of functionality.

Change-Id: I32c4efac2a995352327b8b62b97513cc35a9c5f6