Change return format of UndeletePage::undelete()
- Make it consistently return a StatusValue, as is expected by this
  kind of methods.
- Omit the comment in the returned status, as it matches the one
that was passed as method parameter to undelete().
- Return a good status if there was nothing to undelete. Callers can
  check the status value to determine if something was deleted.

The new format will make getRevisionStatus and getFileStatus
unnecessary, and will also make it impossible to distinguish between
file-related and revision-related errors. I don't think this is really
important (only SpecialUndelete uses the respective methods).

PageArchive is responsible for maintaining backwards compatibility with
the old format.

Bug: T290021
Change-Id: I51179fa7707b5fc44d2257cc927edc0bd798a4d9