resourceloader: use `while` loop in mw.loader#doPropagation()
...instead of `do while`. The difference between them is that
a `while` loop won't run any iterations if the conditions starts
out false, while a `do while` loop will run the first iteration and
only then begin checking the condition. This is useful if you
want to ensure at least one iteration of the loop is run. However,
in this case we know that the condition (`didPropagate`) is
true at the start, because its a variable we just created with
the value `true`. Thus, the `while` loop is also guaranteed to
run at least once. Removing the `do` part saves a few characters
from the output size.

This follows-up 6a83163b1cbad1 where handlePending() was converted
to setAndPropagate(), which removed the need for do-while.

Should be a no-op in terms of functionality.

Change-Id: Ic044fe5afc75131eb0a7bf1cd80e675040be3665