resourceloader: exclude mw.loader test code in production
The line exposing the defineFallbacks() method for tests,
`mw.redefineFallbacksForTest = window.QUnit && defineFallbacks;`,
is run unconditionally on all requests. However, this is only used
for testing in via Special:JavaScriptTest, which is only available
if $wgEnableJavaScriptTest is true, which for production sites is
not the case.* Thus, we don't even need to include the line. Replace
it with a $CODE substitution ResourceLoaderStartUpModule.

Additionally, move it a bit further up in the code to be grouped with
where defineFallbacks() is defined.

Should be a no-op in terms of functionality.

* Technically, we only need it if JavaScript tests are enabled,
and we are on Special:JavaScriptTest, but that doesn't seem
easy to check and isn't as important of an optimization.

Change-Id: I2f0d9443a118e713cc7016c4bc0102fdd2071ec0